+++ src/divide_and_conquer.cpp
@@ -8,7 +8,7 @@
 ClosestPair divide_and_conquer(std::vector<Point2D> points) {
     auto n = points.size();
     ClosestPair closest_pair{};
-    closest_pair.gap = std::numeric_limits<long long>::max();
+    closest_pair.gap = std::numeric_limits<int64_t>::max();
 
     auto temporary_storage(points);
     std::ranges::sort(points);
@@ -42,9 +42,9 @@ ClosestPair divide_and_conquer(std::vector<Point2D> points) {
             y_comparator);
         copy(begin(temporary_storage), begin(temporary_storage) + (r - l), begin(points) + l);
 
-        for (size_t temp_idx{}; auto i : std::views::iota(l, r)) {
+        for (int32_t temp_idx{}; auto i : std::views::iota(l, r)) {
             if (squared_distance(points[i], Point2D{ middle_x, 0 }, Axis::X) >= closest_pair.gap) continue;
-            for (int j = temp_idx - 1;
+            for (int32_t j = temp_idx - 1;
                  j >= 0 && squared_distance(points[i], temporary_storage[j], Axis::Y) < closest_pair.gap;
                  j--) {
                 attempt_to_improve(closest_pair, points[i], temporary_storage[j]);
